handlerとmodel層のテストを分割#9
